add EnemyDestroyedEvent and EnemyDestroyedSystem
