+++ src/events/mod.rs
@@ -1,0 +1,3 @@
+mod events;
+
+pub use self::events::{CollisionEvent, EnemyDestroyedEvent, HitboxCollisionEvent};
